(svn r1631) A couple of fixes for the signed/unsigned warnings. Only train_cmd.c remains to need fixing...

+++ clear_cmd.c
@@ -595,7 +595,7 @@ void TileLoopClearHelper(uint tile)
 		}
 	}
 
-	if (dirty != -1)
+	if (dirty != (uint) -1)
 		MarkTileDirtyByTile(dirty);
 }
